Uber Driver Data Analysis
This Data Set gives the insight of the Rides Undertaken by the Uber Drivers In the USA ,Category of Rides,Average Distance of those rides,Number of trips each hour of the day,Purpose of those rides,Frequencies of Ride on each day of the week,Frequency of ride each day of the month. It also gives insights about the places where the maximum rides were initiated and maximum rides where terminated.